+++ MTP/testing/screens/UIAlert.swift
@@ -1,0 +1,12 @@
+// @copyright Trollwerks Inc.
+
+/// Elements found in alerts
+enum UIAlert: Exposable {
+
+    /// Title
+    case title(String)
+    /// Subtitie
+    case subtitle(String)
+    /// Button
+    case button(String)
+}