+++ CHANGELOG.md
@@ -6,12 +6,20 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+## [Version 1.0.1, Build 639] - 2019-08-20
+
+### Fixed
+- Handle missing locations in ranked user scorecards
+
+## [Version 1.0.1, Build 638] - 2019-08-18
+
 ### Added
 - Photo tag suggestions from location taken
 - Remote notification support
 
 ### Changed
 - Birthdate display removed from My Profile
+- UN visits are checkmarked and have calculation note
 
 ### Fixed
 - UN Country count now synced with location visit changes